Rename paths.py to config.py
