+++ tests/python/test_eval_metrics.py
@@ -6,7 +6,7 @@ import pytest
 import securexgboost as xgb
 import os
 from sklearn.datasets import dump_svmlight_file
-from paths import sym_key_file, priv_key_file, cert_file
+from config import sym_key_file, priv_key_file, cert_file
 
 username = "user1"
 HOME_DIR = os.path.dirname(os.path.realpath(__file__)) + "/../../"
@@ -15,10 +15,6 @@ temp_name_t = HOME_DIR + "demo/data/temp_file_t.txt.train"
 temp_enc_name_t = HOME_DIR + "demo/data/temp_file_t.txt.train.enc"
 temp_name_v = HOME_DIR + "demo/data/temp_file_v.txt.train"
 temp_enc_name_v = HOME_DIR + "demo/data/temp_file_v.txt.train.enc"
-
-#  xgb.init_client(user_name=username, sym_key_file=sym_key_file, priv_key_file=priv_key_file, cert_file=cert_file)
-#  xgb.init_server(enclave_image=HOME_DIR + "build/enclave/xgboost_enclave.signed")
-#  xgb.attest(verify=False)
 
 dpath = HOME_DIR + 'demo/data/'
 rng = np.random.RandomState(1337)
